+++ tests/test_jsonlines/__init__.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-
-sample_text = '{"foo": 1}\n{"ño": 2}\n'
-sample_bytes = sample_text.encode("utf-8")
-sample_data = ({"foo": 1}, {"ño": 2})
-sample_lines = ('{"foo": 1}', '\n', '{"ño": 2}', '\n')