+++ upgrade_readme_usage.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+
 import inspect
 import itertools
 import logging
@@ -13,7 +14,6 @@ import pymince.file
 import pymince.functional
 import pymince.iterator
 import pymince.json
-import pymince.jsonlines
 import pymince.logging
 import pymince.std
 import pymince.text
@@ -29,7 +29,6 @@ modules = (
     pymince.functional,
     pymince.iterator,
     pymince.json,
-    pymince.jsonlines,
     pymince.logging,
     pymince.std,
     pymince.text,