+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## In Progress ##
+## Releases ##
 
 ## v2.9.0 (2023-06-23) ##